requirements.txt
asgiref==3.10.0
base58==2.1.1
boto3==1.40.59
botocore==1.40.59
cffi==2.0.0
cryptography==46.0.3
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jmespath==1.0.1
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
gunicorn
openai==0.28
python-dotenv==1.2.1